Split out result and report data to separate module
